Code cleanup only, no new functions
Remove some logging/trace prints, change int32_t to size_t where appropriate

+++ axolotl/crypto/Ec255PrivateKey.h
@@ -55,7 +55,7 @@ public:
     const std::string serialize() const { return std::string((const char*)keyData_, KEY_LENGTH); }
 
     int32_t getType() const {return EcCurveTypes::Curve25519;}
-    int32_t getEncodedSize() const {return KEY_LENGTH;}
+    size_t getEncodedSize() const {return KEY_LENGTH;}
 
     void getPrivateKey(uint8_t* outBuffer) const {memcpy(outBuffer, keyData_, KEY_LENGTH);}
